feat-#16: move commands to application module
