+++ src/item/application/application.module.ts
@@ -1,4 +1,18 @@
 import { Module } from '@nestjs/common';
+import { CqrsModule } from '@nestjs/cqrs';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-@Module({})
+import { ItemReadRepository, ItemRepository, ItemWriteRepository } from '../repositories';
+import { CommandHandlers } from './commands/handlers';
+
+@Module({
+    imports: [
+        TypeOrmModule.forFeature([ItemReadRepository, ItemWriteRepository]),
+        CqrsModule,
+    ],
+    providers: [
+        ...CommandHandlers,
+        ItemRepository,
+    ]
+})
 export class ApplicationModule {}
