+++ src/item/application/commands/handlers/delete-item-by-id.handler.ts
@@ -2,9 +2,9 @@ import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { DeleteItemByIdCommand } from '../impl';
-import { RpcExceptionService, ErrorValidationService } from '../../../utils';
-import { ItemEntity } from '../../entities';
-import { ItemWriteRepository, ItemRepository } from '../../repositories';
+import { RpcExceptionService, ErrorValidationService } from '../../../../utils';
+import { ItemEntity } from '../../../entities';
+import { ItemWriteRepository, ItemRepository } from '../../../repositories';
 
 @CommandHandler(DeleteItemByIdCommand)
 export class DeleteItemByIdHandler
