+++ src/item/application/commands/impl/create-item.command.ts
@@ -1,4 +1,4 @@
-import { CreateItemDto } from '../../dtos';
+import { CreateItemDto } from '../../../dtos';
 
 export class CreateItemCommand {
   constructor(public readonly createItemDto: CreateItemDto) {}
