+++ src/item/application/commands/impl/delete-item-by-id.command.ts
