+++ src/item/sagas/item.sagas.ts
@@ -3,7 +3,7 @@ import { ICommand, ofType, Saga } from '@nestjs/cqrs';
 import { Observable } from 'rxjs';
 import { delay, map } from 'rxjs/operators';
 
-import { NotifyItemOwnerCommand } from '../commands/impl';
+import { NotifyItemOwnerCommand } from '../application/commands/impl';
 import {
   ItemCreatedEvent,
   ItemUpdatedEvent,
